Refactor basic agent graph types
